Create SlnProjectExtensions and SlnProjectCollectionExtensions per PR comments
